Update Makefile
In an upset of a long standing policy, I am no longer building the objects
in a separate directory, but in the source directory.  GNUMake really
doesn't work "out of the box" to place objects elsewhere so don't even
bother trying anymore.

This does make the Makefile much shorter---the entire library could really
just be:

	libcgi6.a : $(patsubst %.c,%.o,$(shell find src -name '*.c'))
		$(AR) $(ARFLAGS) $@ $?

That's it!  Just those two lines.  Everything else in the Makefile supports
installing, uninstalling and making a distribution tarball.